doc: Documented current models using KDoc syntax
Based on the response received from the request and some information
from the metadata responses (that doesn't match all with the data
responses structure or received values).

+++ preproduction/aemet-open-data-tests/src/main/kotlin/aemet/models/data/AemetRelativeHumidityData.kt
@@ -3,15 +3,50 @@ package org.example.aemet.models.data
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+/**
+ * AEMET relative humidity data.
+ */
 @Serializable
 data class AemetRelativeHumidityData(
+    /**
+     * Max relative humidity.
+     *
+     * @return The percent of relative humidity.
+     */
     @SerialName("maxima") val max: Int,
+
+    /**
+     * Min relative humidity.
+     *
+     * @return The percent of relative humidity.
+     */
     @SerialName("minima") val min: Int,
+
+    /**
+     * Relative humidity levels by hour.
+     */
     @SerialName("dato") val data: List<AemetRelativeHumidityDetailData>,
 )
 
+/**
+ * AEMET relative humidity detail data.
+ */
 @Serializable
 data class AemetRelativeHumidityDetailData(
+    /**
+     * Relative humidity expected percentage value.
+     *
+     * @return Percentage of relative humidity.
+     */
     val value: Int,
-    @SerialName("hora") val hour: Int
+
+    /**
+     * Time of forecast relative humidity.
+     *
+     * @return Possible values for predictions by hour: 0..23.
+     *
+     * Possible values for predictions by days: "00-06",
+     * "06-12", "12-18", "18-24", "00-12", "00-24", "12-24"
+     */
+    @SerialName("hora") val hour: Byte
 )
